CACHE_TIME should be a long value
git-svn-id: https://server.mycore.de/svn/mycore/trunk@29323 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-base/src/main/java/org/mycore/frontend/servlets/MCRServlet.java
@@ -793,7 +793,7 @@ public class MCRServlet extends HttpServlet {
         return MCRTranslation.translate(key, args);
     }
 
-    protected static void writeCacheHeaders(HttpServletResponse response, int CACHE_TIME, long lastModified,
+    protected static void writeCacheHeaders(HttpServletResponse response, long CACHE_TIME, long lastModified,
         boolean useExpire) {
         response.setHeader("Cache-Control", "public, max-age=" + CACHE_TIME);
         response.setDateHeader("Last-Modified", lastModified);
